Binance/requirements.txt
flask
python-binance
pymongo
requests
python-dotenv
